--- src/dtos/monitor-method-update.dto.ts
@@ -1,3 +1,0 @@
-import { MonitorMethodBaseDTO } from './monitor-method-base.dto';
-
-export class UpdateMonitorMethodDTO extends MonitorMethodBaseDTO {}
