--- src/dtos/monitor-plan-comment-update.dto.ts
@@ -1,3 +1,0 @@
-import { MonitorPlanCommentBaseDTO } from './monitor-plan-comment-base.dto';
-
-export class UpdateMonitorPlanCommentDTO extends MonitorPlanCommentBaseDTO {}
