--- src/dtos/unit-capacity-update.dto.ts
@@ -1,3 +1,0 @@
-import { UnitCapacityBaseDTO } from './unit-capacity-base.dto';
-
-export class UpdateUnitCapacityDTO extends UnitCapacityBaseDTO {}
