--- src/dtos/unit-control-update.dto.ts
@@ -1,3 +1,0 @@
-import { UnitControlBaseDTO } from './unit-control-base.dto';
-
-export class UpdateUnitControlDTO extends UnitControlBaseDTO {}
