+++ src/unit-capacity-workspace/unit-capacity.controller.ts
@@ -17,7 +17,7 @@ import { CurrentUser } from '@us-epa-camd/easey-common/decorators';
 import { AuthGuard } from '@us-epa-camd/easey-common/guards';
 import { Logger } from '@us-epa-camd/easey-common/logger';
 
-import { UpdateUnitCapacityDTO } from '../dtos/unit-capacity-update.dto';
+import { UnitCapacityBaseDTO } from '../dtos/unit-capacity.dto';
 import { UnitCapacityDTO } from '../dtos/unit-capacity.dto';
 import { UnitCapacityWorkspaceService } from './unit-capacity.service';
 
@@ -56,7 +56,7 @@ export class UnitCapacityWorkspaceController {
     @CurrentUser() userId: string,
     @Param('locId') locId: string,
     @Param('unitId') unitId: number,
-    @Body() payload: UpdateUnitCapacityDTO,
+    @Body() payload: UnitCapacityBaseDTO,
   ): Promise<UnitCapacityDTO> {
     this.logger.info('Creating Unit Capcity', {
       userId,
@@ -77,7 +77,7 @@ export class UnitCapacityWorkspaceController {
     @Param('locId') locationId: string,
     @Param('unitId') unitId: number,
     @Param('unitCapacityId') unitCapacityId: string,
-    @Body() payload: UpdateUnitCapacityDTO,
+    @Body() payload: UnitCapacityBaseDTO,
     @CurrentUser() userId: string,
   ): Promise<UnitCapacityDTO> {
     this.logger.info('Updating Unit Capacity', {
